Move checkout promo code tests to new file
